headers(fix): move components to specific folder
