+++ src/pages/lobby.tsx
@@ -15,7 +15,7 @@ import ROUTES from '../routes'
 import { useContributionStore, Store } from '../store/contribute'
 import { isSuccessRes, sleep } from '../utils'
 import { useAuthStore } from '../store/auth'
-import HeaderJustGoingBack from '../components/HeaderJustGoingBack'
+import HeaderJustGoingBack from '../components/headers/HeaderJustGoingBack'
 import useSequencerStatus from '../hooks/useSequencerStatus'
 import { Trans, useTranslation } from 'react-i18next'
 import { ErrorRes } from '../types'
